pyproject.toml
[tool.poetry]
name = "biot"
version = "0.1.0"
description = ""
authors = ["Bob Hong <bhong@plex.com>"]

[tool.poetry.dependencies]
python = "^3.9"
Faker = "^13.13.0"
paho-mqtt = "^1.6.1"
python-dotenv = "^0.20.0"
influxdb-client = "^1.29.1"
lxml = "^4.9.0"
requests = "^2.28.0"

[tool.poetry.dev-dependencies]
autopep8 = "^1.6.0"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
